dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile dev-requirements.in
#
black==23.1.0
    # via -r dev-requirements.in
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   black
    #   flask
    #   functions-framework
cloudevents==1.9.0
    # via functions-framework
deprecation==2.1.0
    # via cloudevents
flake8==6.0.0
    # via -r dev-requirements.in
flask==2.2.3
    # via functions-framework
functions-framework==3.3.0
    # via -r dev-requirements.in
google-api-core[grpc]==2.11.0
    # via
    #   google-api-python-client
    #   google-cloud-pubsub
google-api-python-client==2.79.0
    # via -r dev-requirements.in
google-auth==2.16.1
    # via
    #   -r dev-requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-pubsub==2.15.0
    # via -r dev-requirements.in
googleapis-common-protos[grpc]==1.58.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via google-cloud-pubsub
grpcio==1.51.3
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.51.3
    # via
    #   google-api-core
    #   google-cloud-pubsub
gunicorn==20.1.0
    # via functions-framework
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.0
    # via
    #   black
    #   deprecation
pathspec==0.11.0
    # via black
platformdirs==3.0.0
    # via black
proto-plus==1.22.2
    # via google-cloud-pubsub
protobuf==4.22.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyparsing==3.0.9
    # via httplib2
python-dotenv==1.0.0
    # via -r dev-requirements.in
requests==2.28.2
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
tomli==2.0.1
    # via black
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.14
    # via requests
watchdog==2.3.0
    # via functions-framework
werkzeug==2.2.3
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
